Move base post class to helper module
